+++ src/backend/ci/core/common/common-api/src/main/kotlin/com/tencent/devops/common/api/util/DateTimeUtil.kt
@@ -26,6 +26,7 @@
 
 package com.tencent.devops.common.api.util
 
+import org.slf4j.LoggerFactory
 import java.text.SimpleDateFormat
 import java.time.LocalDateTime
 import java.time.ZoneId
@@ -49,6 +50,27 @@ fun LocalDateTime.timestampmilli(): Long {
 }
 
 object DateTimeUtil {
+
+    private val formatter = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ")
+
+    private val logger = LoggerFactory.getLogger(DateTimeUtil::class.java)
+
+    /**
+     * 单位转换，分钟转换秒
+     */
+    fun minuteToSecond(minutes: Int): Int {
+        return minutes * 60
+    }
+
+    /**
+     * 单位转化，秒转换分钟
+     * 以Int为计算单位，有余数将省去
+     */
+    fun secondToMinute(seconds: Int): Int {
+        return seconds / 60
+    }
+
+
     /**
      * 获取从当前开始一定单位时间间隔的日期
      * @param unit 单位 Calendar.SECONDS
@@ -85,6 +107,20 @@ object DateTimeUtil {
         val instant = dateTime.atZone(zone).toInstant()
         val simpleDateFormat = SimpleDateFormat(format)
         return simpleDateFormat.format(Date.from(instant))
+    }
+
+    /**
+     * 单位转化，秒转换时间戳
+     * 2019-09-02T08:58:46+0000 -> xxxxx
+     */
+    fun zoneDateToTimestamp(timeStr: String?): Long {
+        try {
+            if (timeStr.isNullOrBlank()) return 0L
+            return formatter.parse(timeStr).time
+        } catch (e: Exception) {
+            logger.error("fail to parse time string: $timeStr", e)
+        }
+        return 0L
     }
 
     /**
